Add reuseable stories command to project slash command
